Add game creation and joining mechanics
This makes the main page a place which can direct users to the create game
page, or tells them to paste a link to a game page. Games are created with
blind which must be divisble by two, a starting chip count, and a passphrase so
strangers cannot join. The server assigns a gameID to a newly created game.
When joining a game you just need to paste your url and enter the passphrase
for the game

Notes:
* Using mutex to lock games instead of channel. Some golang wiki said it was
more idiomatic for managing state.
* Not using any framework for links in react app, letting MainComponent manage
history and urls